feat: made sure not to excessively look back in the reward function inside BucketSubmission.sol

+++ src/temp/MD2.sol
@@ -3,9 +3,9 @@ pragma solidity 0.8.21;
 
 import "forge-std/console.sol";
 import "@/MinerPoolAndGCA/BucketSubmission.sol";
+
 contract MD2 is BucketSubmission {
-    
-    function addToCurrentBucket(address grcToken,uint amount) external {
-        _addToCurrentBucket(grcToken,amount);
+    function addToCurrentBucket(address grcToken, uint256 amount) external {
+        _addToCurrentBucket(grcToken, amount);
     }
 }
